Fix #224
There is a general problem handling dates before
year 1000. They are not sorted will and an error
will occure if the date is BC and converted with
xsd:date.

This fix will solve the problem for the topic
aspect, but there may be other problems. An
author such as Plato does not seem to render ok.