Fix #262 for topics and implement topic map
An attempt to fix the memory problem with big graphs #262
has been made by limiting the number of links returned to
400. There might be other panels where there is a similar
problem.

Co-occurring topics map has now also been implemented.
For some topics there might not be relevant co-occuring
topics that are geolocatable and that means that the
map will be grey and empty.