Fix publication panel in chemical aspect
A pull request with commit https://github.com/fnielsen/scholia/pull/950
introduced and error in the SPARQL for the "Recently published works on the
chemical" so it did not show anything. This was noted by @Daniel-Mietchen.
@egonw believed this to be an server side error, but it seem that some
extra erroneous code was pasted(?) into a subexpression.
An issue was filed for what was thought to be a server-side problem at #994.
This commit should fix #994.