scholia/tex.py: Do not escape bibtex key
When writing a bib-entry, the key must not be escaped, otherwise
bibtex will not find the key given in the tex file in the bib file.

One example where this breaks without this change is the DOI
10.1007/11925941_2 (due to the underscore). Underscores are special
characters in LaTeX, but not in Bibtex keys, and are allowed in
DOIs.