Fix #1459: venue panel distinct authors
When publications had multiple publication dates an author would
appear multiple times in the "recently published works" panel in the
venue aspect. With a distinct an author now only is listed a single time.

This might cause confusion if there are multiple different authors with the
same name. Alternatively the SPARQL should be further refined.