Fix #1736: Tox error and style error in #1722
The #1722 pull request introduced a couple of minor styling errors as well
as an incompatibility with Python 2 and since this was tested via the tox.ini
file, a text with tox would err.

This PR fixes the two styling errors and also more importantly removes the
test with Python 2.7, so that Python 2.x is no longer supported. The default
Python installation on older Linux system, e.g., Ubuntu 18.04 LTS would then
not work. python3 in such systems which uses 3.6 should now still work,
thus Scholia should be started with python3 runserver.py on such old systems.