Fix #2178, #2168 and #2177 in quickstatement
In OJS scrape, and and possibly other cases, strings to quickstatements may
have a tab or trailing spaces that will confuse quickstatements. This commit
normalize the string for title and author before generating the quickstatement
line. Furthermore if the string contains double quotes the previous code
generated wrong quickstatements with a wrong kind of escape. No escape seems to
be necessary and it has been removed.