Fix #2534 and #2535 Test in wikipedia and arxiv
Doctest in arxiv.py failed because of change in error message from the arxiv
server. A more general test is now done in the doctest which is not dependent
on the specicific error message.

A doctest in wikipedia.py failed because the default requests user agent was
banned by the Wikidata Query Service. Now the user agent is set based on the
Scholia configuration.