Fix #2673 Use API instead of SPARQL for redirect
Redirect in the Scholia web app used SPARQL to get 'instance of'
information for choosing the appropriate aspect. There has
recently been Internal Server Error associated with this use it seems.

Now the download of the 'instance of' happens via the API. Whether
this approach helps on the fundamental problem is not clear.
It is difficult to test as the problem only occures on the Toolforge
version of Scholia, - not locally.

The long 'if'-sentence for aspect guessing is now copied to 'views.py'
which seems to be a more appropriate place.

The new approach is not so general as the SPARQL-based approach as some
selection was based on a '+' property search.

This commit also include a fix for the CLI on search that did not work.